refactor(roles): create unified roles module and delete manager directory
Create roles/ module as single source of truth for role
definitions, request builders, and role-specific configs.
Delete manager/ directory and redundant execution files.

Related: issue-460
