+++ src/vibe3/domain/handlers/issue_state_dispatch.py
@@ -1,4 +1,4 @@
-"""Event handlers for manager dispatch."""
+"""Generic IssueStateChanged -> role dispatch handler."""
 
 import asyncio
 from typing import Callable
@@ -7,54 +7,46 @@ from loguru import logger
 
 from vibe3.domain.events import DomainEvent
 from vibe3.domain.events.flow_lifecycle import IssueStateChanged
-from vibe3.execution.role_services import (
-    build_manager_dispatch_request,
-    resolve_orchestra_repo_root,
-)
 from vibe3.models.orchestra_config import OrchestraConfig
 from vibe3.models.orchestration import IssueInfo, IssueState
+from vibe3.roles.registry import build_issue_state_request, resolve_issue_state_role
 
-_MANAGER_TRIGGER_STATES: frozenset[str] = frozenset({"ready", "handoff"})
 
+def handle_issue_state_changed_for_roles(event: IssueStateChanged) -> None:
+    """Dispatch triggerable roles for IssueStateChanged events.
 
-def handle_issue_state_changed_for_manager(event: IssueStateChanged) -> None:
-    """Dispatch manager when an issue enters a manager-consumable state.
-
-    Schedules the actual dispatch as an async task on the running event loop
-    to avoid blocking the loop with synchronous I/O (same pattern as
-    handle_governance_scan_started).
-
-    Uses issue_title carried in the event payload (populated by OrchestrationFacade
-    from the polling IssueInfo) to avoid a redundant view_issue GitHub API call.
-    Falls back to fetching from GitHub only when the title is absent (webhook path).
+    Current supported issue-state role:
+    - manager (ready / handoff)
     """
-    if event.to_state not in _MANAGER_TRIGGER_STATES:
+    role = resolve_issue_state_role(event.to_state)
+    if role is None:
         return
 
     logger.bind(
-        domain="manager_handler",
+        domain="issue_state_dispatch_handler",
+        role=role.registry_role,
         issue_number=event.issue_number,
         from_state=event.from_state,
         to_state=event.to_state,
-    ).info("Manager handler triggered, scheduling async dispatch")
+    ).info("Issue-state role handler triggered, scheduling async dispatch")
 
     async def _do_dispatch() -> None:
         loop = asyncio.get_event_loop()
         config = OrchestraConfig.from_settings()
 
         target_state = (
-            IssueState.READY if event.to_state == "ready" else IssueState.HANDOFF
+            IssueState.READY
+            if event.to_state == IssueState.READY.value
+            else IssueState.HANDOFF
         )
 
         if event.issue_title is not None:
-            # Fast path: polling events carry the title → no GitHub API call needed.
             issue_info: IssueInfo | None = IssueInfo(
                 number=event.issue_number,
                 title=event.issue_title,
                 state=target_state,
             )
         else:
-            # Slow path: webhook events don't carry the title → fetch from GitHub.
             from vibe3.clients.github_client import GitHubClient
 
             github_client = GitHubClient()
@@ -64,7 +56,8 @@ def handle_issue_state_changed_for_manager(event: IssueStateChanged) -> None:
 
             if issue_data is None or isinstance(issue_data, str):
                 logger.bind(
-                    domain="manager_handler",
+                    domain="issue_state_dispatch_handler",
+                    role=role.registry_role,
                     issue_number=event.issue_number,
                     error="issue_not_found",
                 ).error("Failed to fetch issue details from GitHub")
@@ -73,7 +66,8 @@ def handle_issue_state_changed_for_manager(event: IssueStateChanged) -> None:
             issue_info = IssueInfo.from_github_payload(issue_data)
             if issue_info is None:
                 logger.bind(
-                    domain="manager_handler",
+                    domain="issue_state_dispatch_handler",
+                    role=role.registry_role,
                     issue_number=event.issue_number,
                     error="invalid_issue_data",
                 ).error("Failed to parse issue data from GitHub response")
@@ -83,12 +77,12 @@ def handle_issue_state_changed_for_manager(event: IssueStateChanged) -> None:
 
         if issue_info is None:
             logger.bind(
-                domain="manager_handler",
+                domain="issue_state_dispatch_handler",
+                role=role.registry_role,
                 issue_number=event.issue_number,
-            ).error("Issue info is None, cannot dispatch manager")
+            ).error("Issue info is None, cannot dispatch role")
             return
 
-        # Use unified ExecutionCoordinator path
         from vibe3.agents.backends.codeagent import CodeagentBackend
         from vibe3.clients.sqlite_client import SQLiteClient
         from vibe3.environment.session_registry import SessionRegistryService
@@ -100,68 +94,67 @@ def handle_issue_state_changed_for_manager(event: IssueStateChanged) -> None:
         coordinator = ExecutionCoordinator(config, store)
 
         try:
-            # Prepare manager-specific request
             request = await loop.run_in_executor(
                 None,
-                lambda: build_manager_dispatch_request(
+                lambda: build_issue_state_request(
                     config,
                     issue_info,
+                    event.to_state,
                     registry=registry,
-                    repo_path=resolve_orchestra_repo_root(),
                 ),
             )
 
             if request is None:
                 logger.bind(
-                    domain="manager_handler",
+                    domain="issue_state_dispatch_handler",
+                    role=role.registry_role,
                     issue_number=event.issue_number,
-                ).error("Failed to prepare manager execution request")
+                ).error("Failed to prepare role execution request")
                 return
 
-            # Dispatch via unified coordinator
             result = await loop.run_in_executor(
                 None, lambda: coordinator.dispatch_execution(request)
             )
 
             if result.launched:
                 logger.bind(
-                    domain="manager_handler",
+                    domain="issue_state_dispatch_handler",
+                    role=role.registry_role,
                     issue_number=event.issue_number,
-                ).success("Manager execution launched via ExecutionCoordinator")
+                ).success("Role execution launched via ExecutionCoordinator")
             else:
                 logger.bind(
-                    domain="manager_handler",
+                    domain="issue_state_dispatch_handler",
+                    role=role.registry_role,
                     issue_number=event.issue_number,
-                ).warning(f"Manager dispatch failed: {result.reason}")
+                ).warning(f"Role dispatch failed: {result.reason}")
 
         except Exception as exc:
             logger.bind(
-                domain="manager_handler",
+                domain="issue_state_dispatch_handler",
+                role=role.registry_role,
                 issue_number=event.issue_number,
-            ).exception(f"Manager dispatch failed: {exc}")
+            ).exception(f"Role dispatch failed: {exc}")
 
     try:
-        # Called from within heartbeat's async event loop — schedule as task.
         loop = asyncio.get_running_loop()
         loop.create_task(
             _do_dispatch(),
-            name=f"manager-dispatch-{event.issue_number}-{event.to_state}",
+            name=f"issue-state-role-dispatch-{event.issue_number}-{event.to_state}",
         )
     except RuntimeError:
-        # No running loop (e.g. tests, direct CLI call) — safe to use asyncio.run().
         asyncio.run(_do_dispatch())
 
 
-def register_manager_handlers() -> None:
-    """Register all manager event handlers."""
+def register_issue_state_dispatch_handlers() -> None:
+    """Register generic IssueStateChanged role dispatch handlers."""
     from typing import cast
 
     from vibe3.domain.publisher import subscribe
 
-    # Subscribe to IssueStateChanged for manager trigger
     subscribe(
         "IssueStateChanged",
-        cast(Callable[[DomainEvent], None], handle_issue_state_changed_for_manager),
+        cast(Callable[[DomainEvent], None], handle_issue_state_changed_for_roles),
     )
 
-    logger.bind(domain="events").info("Manager event handlers registered")
+    logger.bind(domain="events").info("Issue-state role dispatch handlers registered")
