+++ tests/vibe3/commands/test_internal.py
@@ -11,21 +11,24 @@ runner = CliRunner()
 
 def test_internal_manager_dispatch():
     """测试 internal manager 命令参数解析和调用."""
-    with patch("vibe3.manager.manager_run_service.run_manager_issue_mode") as mock_run:
+    with patch(
+        "vibe3.execution.issue_role_sync_runner.run_issue_role_mode"
+    ) as mock_run:
         result = runner.invoke(cli_app, ["internal", "manager", "123", "--no-async"])
 
         assert result.exit_code == 0
-        mock_run.assert_called_once_with(
-            issue_number=123,
-            dry_run=False,
-            async_mode=False,  # --no-async
-            fresh_session=False,
-        )
+        assert mock_run.call_args.kwargs["issue_number"] == 123
+        assert mock_run.call_args.kwargs["dry_run"] is False
+        assert mock_run.call_args.kwargs["async_mode"] is False
+        assert mock_run.call_args.kwargs["fresh_session"] is False
+        assert mock_run.call_args.kwargs["spec"].role_name == "manager"
 
 
 def test_internal_manager_with_options():
     """测试 internal manager 命令支持所有选项."""
-    with patch("vibe3.manager.manager_run_service.run_manager_issue_mode") as mock_run:
+    with patch(
+        "vibe3.execution.issue_role_sync_runner.run_issue_role_mode"
+    ) as mock_run:
         result = runner.invoke(
             cli_app,
             [
@@ -39,12 +42,11 @@ def test_internal_manager_with_options():
         )
 
         assert result.exit_code == 0
-        mock_run.assert_called_once_with(
-            issue_number=456,
-            dry_run=True,
-            async_mode=False,
-            fresh_session=True,
-        )
+        assert mock_run.call_args.kwargs["issue_number"] == 456
+        assert mock_run.call_args.kwargs["dry_run"] is True
+        assert mock_run.call_args.kwargs["async_mode"] is False
+        assert mock_run.call_args.kwargs["fresh_session"] is True
+        assert mock_run.call_args.kwargs["spec"].role_name == "manager"
 
 
 def test_internal_apply_dispatch():
